Introduce fast switchovers/failovers
* Introduce customer configurable topology_query. In order to get the
topology from a cluster, the query should be
`topology_query='select endpoint from rds_topology'`. If this query is not
set, then fast switchovers will not be enabled for the database.
* The topology_query is run every time PgBouncer is started as well as reloaded
to discover all nodes in a cluster.
* Once the topology_query runs, connection pools are created to all nodes in a
cluster. A query of select pg_is_in_recovery() is run by the proxy to discover
who the writer in the cluster is and caches this node.
* A {switch,fail}over is discovered when the connection to the original writer are
killed. PgBouncer pauses the client and serially polls the remaining nodes with
`pg_is_in_recovery()` to discover the new writer. The new writer is cached for future
connections.
* Introduced a config option, `server_failed_delay`, which specifies how long to wait
between failed connection attempts to the downed writer to attempt a reconnect. The
default is 30 seconds. This is to prevent infinite polling on the client.
* Once the downed writer is detected as active, a new connection pool is created to keep
it available for the next switchover.
* If using fast switchovers, PgBouncer will override the server_lifetime option to never
allow closing existing connections. This allows the connection pools to remain open indefinitely.
* A config option (polling_frequency) to determine how frequently to poll nodes during
a {switch,fail}over has been added. Defaults to 100ms.
* Change server_check_delay to be set to 0 if fast_switchovers are
enabled. This is to allow the polling logic to take place for fast
switchovers.