Bug fixes for fast switchovers and change in topology_query behavior
- Allow topology as long as there are at least two nodes
It was on oversight by me to enforce 3 nodes in the cluster. What if we
use this for MAZ in the future or for a cluster with N nodes?

- Send the topology query every time a switchover occurs
We use this to ensure we get a better accounting of the PgBouncer usage
metric.

- Remove global checking for new writer
This fixes the theoretical case where one DB is switching over, but
others are operating fine. Setting `checking_for_new_writer` globally
would block all DBs. Instead, we want to only block the DB with pools
that are switching over.

-  Fix backwards logic in check_unused_servers
Found this when running the in-repo regression tests. If the
topology_query is not set, then continue with the other checks in the
function. The original check was ensuring that pool->db was not NULL
before accessing topology_query.